Feat: 뷰와 뷰 컨트롤러 분리
